Made splitting MX nodes to categories by nodegroup
