+++ src/hdusd/mx_nodes/nodes/__init__.py
@@ -14,11 +14,11 @@
 # ********************************************************************
 import bpy
 import nodeitems_utils
-from nodeitems_utils import NodeCategory, NodeItem
 
 from ...utils import HDUSD_LIBS_DIR
 from .. import log
 from .base_node import create_node_types, MxNodeSocket
+from . import categories
 
 
 node_types = create_node_types([
@@ -28,32 +28,20 @@ node_types = create_node_types([
 ])
 
 
-class MxNodeCategory(NodeCategory):
-    @classmethod
-    def poll(cls, context):
-        return context.space_data.tree_type == 'hdusd.MxNodeTree'
-
-
-node_categories = [
-    MxNodeCategory('HdUSD_MX_SHADERS', "Shaders", items=[
-        NodeItem(node_type.bl_idname) for node_type in node_types
-    ]),
-]
-
-# nodes to register
-register_nodes, unregister_nodes = bpy.utils.register_classes_factory(node_types)
 register_sockets, unregister_sockets = bpy.utils.register_classes_factory([
     MxNodeSocket,
 ])
+register_nodes, unregister_nodes = bpy.utils.register_classes_factory(node_types)
 
 
 def register():
     register_sockets()
     register_nodes()
-    nodeitems_utils.register_node_categories("MX_NODES", node_categories)
+
+    nodeitems_utils.register_node_categories("'HdUSD_MX_NODES", categories.get_node_categories())
 
 
 def unregister():
-    nodeitems_utils.unregister_node_categories("MX_NODES")
+    nodeitems_utils.unregister_node_categories("'HdUSD_MX_NODES")
     unregister_nodes()
     unregister_sockets()
